+++ src/Tests/Functions.fs
@@ -440,4 +440,63 @@ let tests = testList "Functions tests" [
         let vnetConnections = resources |> getResource<Web.VirtualNetworkConnection> 
         Expect.hasLength vnetConnections 1 "incorrect number of Vnet connections"
     }
+
+    test "Supports slot settings" {
+        let functionsApp = functions { name "test"; slot_settings [ "sticky_config", "sticky_config_value"; "another_sticky_config", "another_sticky_config_value" ]} 
+
+        let scn = functionsApp |> getResources |> getResource<Web.SlotConfigName> |> List.head
+        let ws = functionsApp |> getResources |> getResource<Web.Site> |> List.head
+
+        let template = arm{ add_resource functionsApp}
+        let jobj = template.Template |> Writer.toJson |> Newtonsoft.Json.Linq.JObject.Parse
+
+        let appSettingNames = 
+            jobj.SelectTokens $"$..resources[?(@name=='{functionsApp.Name.ResourceName.Value}/slotconfignames')].properties.appSettingNames[*]" 
+            |> Seq.map (fun x -> x.ToString())
+
+        let dependencies = 
+            jobj.SelectTokens $"$..resources[?(@name=='{functionsApp.Name.ResourceName.Value}/slotconfignames')].dependsOn[*]" 
+            |> Seq.map (fun x -> x.ToString())
+
+        let expectedSettings = Map [ 
+            "sticky_config", LiteralSetting "sticky_config_value"
+            "another_sticky_config", LiteralSetting "another_sticky_config_value" ]
+              
+        let settings = Expect.wantSome ws.AppSettings "AppSettings should be set"
+        Expect.containsAll  settings  expectedSettings "App settings should contain the slot settings"
+        Expect.containsAll scn.SlotSettingNames ["sticky_config"; "another_sticky_config"] "Slot config names should be set"
+        Expect.equal scn.SiteName (ResourceName "test") "Parent name should be set"
+        Expect.containsAll appSettingNames  [ "sticky_config"; "another_sticky_config"] "Slot config names should be present in template"
+        Expect.containsAll dependencies  [ $"[resourceId('Microsoft.Web/sites', '{functionsApp.Name.ResourceName.Value}')]"] "Slot config names resource should depend on web site"
+
+    }
+
+    test "Supports slot setting" {
+        let functionsApp = functions { name "test"; slot_setting "sticky_config" "sticky_config_value" } 
+
+        let scn = functionsApp |> getResources |> getResource<Web.SlotConfigName> |> List.head
+        let ws = functionsApp |> getResources |> getResource<Web.Site> |> List.head
+
+        let template = arm{ add_resource functionsApp}
+        let jobj = template.Template |> Writer.toJson |> Newtonsoft.Json.Linq.JObject.Parse
+
+        let appSettingNames = 
+            jobj.SelectTokens $"$..resources[?(@name=='{functionsApp.Name.ResourceName.Value}/slotconfignames')].properties.appSettingNames[*]" 
+            |> Seq.map (fun x -> x.ToString())
+
+        let dependencies = 
+            jobj.SelectTokens $"$..resources[?(@name=='{functionsApp.Name.ResourceName.Value}/slotconfignames')].dependsOn[*]" 
+            |> Seq.map (fun x -> x.ToString())
+
+        let expectedSettings = Map [ 
+            "sticky_config", LiteralSetting "sticky_config_value" ]
+                
+        let settings = Expect.wantSome ws.AppSettings "AppSettings should be set"
+        Expect.containsAll  settings  expectedSettings "App settings should contain the slot setting"
+        Expect.containsAll scn.SlotSettingNames ["sticky_config"] "Slot config name should be set"
+        Expect.equal scn.SiteName (ResourceName "test") "Parent name should be set"
+        Expect.containsAll appSettingNames  [ "sticky_config" ] "Slot config name should be present in template"
+        Expect.containsAll dependencies  [ $"[resourceId('Microsoft.Web/sites', '{functionsApp.Name.ResourceName.Value}')]"] "Slot config names resource should depend on web site"
+
+    }
 ]